fix[g10]: fixed code as asked.. deleted mapObjectRepository class and removed service and repository propery from mainWindow
