+++ HealthcareSystem/GraphicalEditor/Models/MapObjectRelated/MapObject.cs
@@ -31,6 +31,8 @@ namespace GraphicalEditor.Models
             }
         }
 
+        public Line Line { get; private set; }
+
         public MapObject(MapObjectEntity mapObjectEntity, MapObjectMetrics mapObjectMetrics, MapObjectDoor mapObjectDoor)
         {
             MapObjectEntity = mapObjectEntity;
@@ -40,8 +42,21 @@ namespace GraphicalEditor.Models
             MapObjectDoor.MapObjectMetrics = mapObjectMetrics;
 
             RectangleInitialization();
+            if (mapObjectEntity.MapObjectType.TypeOfMapObject == TypeOfMapObject.ROAD) {
+                LineInitialization();
+            }
             PositionObjectNameTextBlock();
         }
+        private void LineInitialization()
+        {
+            Line = new Line();
+            Line.Fill = Brushes.White;
+            Line.X1 = Rectangle.RadiusX;
+            Line.X2 = Rectangle.Width+ Rectangle.RadiusX;
+            Line.Y1 = Rectangle.Height / 2-2;
+            Line.Y2 = Rectangle.Height / 2 + 2;
+        }
+
 
         private void RectangleInitialization()
         {
@@ -49,9 +64,7 @@ namespace GraphicalEditor.Models
             Rectangle.Fill = MapObjectEntity.ObjectEntityColor;
             Rectangle.Height = MapObjectMetrics.HeightOfMapObject;
             Rectangle.Width = MapObjectMetrics.WidthOfMapObject;
-
             MapObjectEntity.SetStrokeAndStrokeThickness(Rectangle);
-
         }
 
         private void PositionObjectNameTextBlock()
@@ -74,6 +87,9 @@ namespace GraphicalEditor.Models
             Canvas.SetTop(Rectangle, MapObjectMetrics.YOfCanvas);
             canvas.Children.Add(MapObjectDoor.GetDoor());
             canvas.Children.Add(MapObjectNameTextBlock);
+            if (MapObjectEntity.MapObjectType.TypeOfMapObject == TypeOfMapObject.ROAD) {
+                canvas.Children.Add(Line);
+            }
         }
     }
 }
