+++ HealthcareSystem/GraphicalEditor/Models/MapObjectRelated/Road.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Shapes;
-using System.Windows.Media;
 
 namespace GraphicalEditor.Models.MapObjectRelated
 {
@@ -16,9 +15,12 @@ namespace GraphicalEditor.Models.MapObjectRelated
         {
             FormatObjectDescription(Description);
         }
-        
+
+        public Line Line{get; set;}
+
         public override void FormatObjectDescription(string description)
         {
+
             if (String.IsNullOrEmpty(description))
             {
                 Description = MapObjectType.ObjectTypeFullName;
