+++ HealthcareSystem/GraphicalEditor/Repository/FileRepository.cs
@@ -47,6 +47,36 @@ namespace GraphicalEditor.Repository
             {
                 return new List<MapObject>();
             }
-        }        
+        }
+        public void UpdateMapObject(MapObject mapObject)
+        {
+            IEnumerable<MapObject> allMapObjects = LoadMap();
+            var singleMapObject = allMapObjects.FirstOrDefault(e => e.MapObjectEntity.Id == mapObject.MapObjectEntity.Id);
+            if (singleMapObject != null)
+            {
+                List<MapObject> objectToSave = new List<MapObject>();
+                foreach (MapObject anObject in allMapObjects)
+                {
+                    objectToSave.Add(anObject);
+
+                }
+
+                for (int i = 0; i < objectToSave.Count; i++)
+                {
+
+                    if (objectToSave[i].MapObjectEntity.Id == mapObject.MapObjectEntity.Id)
+                    {
+                        objectToSave[i].MapObjectEntity.MapObjectType = mapObject.MapObjectEntity.MapObjectType;
+                        objectToSave[i].MapObjectEntity.Description = mapObject.MapObjectEntity.Description;
+                        objectToSave[i].MapObjectMetrics = mapObject.MapObjectMetrics;
+                        objectToSave[i].Rectangle.Fill = mapObject.Rectangle.Fill;
+                        objectToSave[i].MapObjectDoor = mapObject.MapObjectDoor;
+                        objectToSave[i].MapObjectNameTextBlock = mapObject.MapObjectNameTextBlock;
+                        break;
+                    }
+                }
+                SaveMap(objectToSave);
+            }
+        }
     }
 }
