--- HealthcareSystem/GraphicalEditor/Repository/Interface/IMapObjectRepository.cs
@@ -1,14 +1,0 @@
-﻿using GraphicalEditor.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GraphicalEditor.Repository.Intefrace
-{
-    public interface IMapObjectRepository
-    {
-        void UpdateMapObject(MapObject mapObject);
-    }
-}
